# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification - Transporte Escolar Eventos
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /escolas:
    get:
      operationId: listEscolas
      tags: [escolas]
      summary: Listar escolas
      responses:
        "200":
          description: Lista de escolas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Escola"
    post:
      operationId: createEscola
      tags: [escolas]
      summary: Cadastrar escola
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EscolaInput"
      responses:
        "201":
          description: Escola criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Escola"
  /escolas/{id}:
    get:
      operationId: getEscola
      tags: [escolas]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Escola
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Escola"
    patch:
      operationId: updateEscola
      tags: [escolas]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EscolaUpdate"
      responses:
        "200":
          description: Escola atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Escola"
    delete:
      operationId: deleteEscola
      tags: [escolas]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deletado

  # ── Motoristas ──────────────────────────────────────────────────────────────
...
  /escalas:
    get:
      operationId: listEscalas
      tags: [escalas]
      summary: Listar escalas de transporte
      parameters:
        - name: sessao_id
          in: query
          schema: { type: integer }
        - name: escola_id
          in: query
          schema: { type: integer }
        - name: evento_id
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Lista de escalas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Escala"
    post:
      operationId: createEscala
      tags: [escalas]
      summary: Criar escala de transporte
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EscalaInput"
      responses:
        "201":
          description: Escala criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Escala"
  /escalas/{id}:
    get:
      operationId: getEscala
      tags: [escalas]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Escala
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Escala"
    patch:
      operationId: updateEscala
      tags: [escalas]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EscalaUpdate"
      responses:
        "200":
          description: Escala atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Escala"
    delete:
      operationId: deleteEscala
      tags: [escalas]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deletado
  /escalas/{id}/capacidade:
    get:
      operationId: getEscalaCapacidade
      tags: [escalas]
      summary: Análise de capacidade de uma escala
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Análise de capacidade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EscalaCapacidade"
components:
  schemas:
...
    Escola:
      type: object
      required: [id, nome]
      properties:
        id:          { type: integer }
        nome:        { type: string }
        num_prodesp: { type: ["string", "null"] }
        endereco:    { type: ["string", "null"] }
        cidade:      { type: ["string", "null"] }
        total_alunos: { type: ["integer", "null"] }
    EscolaInput:
      type: object
      required: [nome]
      properties:
        nome:        { type: string }
        num_prodesp: { type: string }
        endereco:    { type: string }
        cidade:      { type: string }
    EscolaUpdate:
      type: object
      properties:
        nome:        { type: string }
        num_prodesp: { type: string }
        endereco:    { type: string }
        cidade:      { type: string }

    # ── Motorista ─────────────────────────────────────────────────────────────
...
    Escala:
      type: object
      required: [id, sessao_id, escola_id, veiculo_id]
      properties:
        id:                  { type: integer }
        sessao_id:           { type: integer }
        escola_id:           { type: integer }
        escola_nome:         { type: ["string", "null"] }
        veiculo_id:          { type: integer }
        veiculo_placa:       { type: ["string", "null"] }
        veiculo_modelo:      { type: ["string", "null"] }
        motorista_id:        { type: ["integer", "null"] }
        motorista_nome:      { type: ["string", "null"] }
        vagas_disponibilizadas: { type: integer }
        horario_embarque:    { type: ["string", "null"] }
        periodo:             { type: ["string", "null"] }
        segmento_filtro:     { type: ["string", "null"] }
        turma_filtro:        { type: ["string", "null"] }
        total_alunos:        { type: ["integer", "null"] }
    EscalaInput:
      type: object
      required: [sessao_id, escola_id, veiculo_id]
      properties:
        sessao_id:              { type: integer }
        escola_id:              { type: integer }
        veiculo_id:             { type: integer }
        motorista_id:           { type: integer }
        vagas_disponibilizadas: { type: integer }
        periodo:                { type: string }
        segmento_filtro:        { type: string }
        turma_filtro:           { type: string }
    EscalaUpdate:
      type: object
      properties:
        veiculo_id:             { type: integer }
        motorista_id:           { type: integer }
        vagas_disponibilizadas: { type: integer }
        periodo:                { type: string }
        segmento_filtro:        { type: string }
        turma_filtro:           { type: string }
    EscalaCapacidade:
      type: object
      required: [escala_id, total_alunos, total_vagas, vagas_sobra, horario_embarque]
      properties:
        escala_id:       { type: integer }
        escola_nome:     { type: string }
        total_alunos:    { type: integer }
        total_vagas:     { type: integer }
        vagas_sobra:     { type: integer }
        horario_embarque: { type: ["string", "null"] }
        tempo_deslocamento_minutos: { type: ["integer", "null"] }

    # ── Importação ────────────────────────────────────────────────────────────
...