# lib/api-zod/src/generated/types/escolaUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification - Transporte Escolar Eventos
 * OpenAPI spec version: 0.1.0
 */

export interface EscolaUpdate {
  nome?: string;
  num_prodesp?: string;
  endereco?: string;
  cidade?: string;
}